Add "Extractor" to variable `site_name`
The docs in this repo focus on the extractor, they don't explain the entirety of the newpipe codebase. This should be made clear everywhere it isn't, starting with the site name, so prospective contributors coming for an explanation of the architecture of the newpipe app's repo don't waste their time.